pyproject.toml
[tool.poetry]
name = "myFlask"
version = "0.1.0"
description = ""
authors = ["gwm <403052870@qq.com>"]

[[tool.poetry.source]]
name = "douban"
url = "https://pypi.doubanio.com/simple/"
default = true

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
flask-sqlalchemy = "^2.4.4"
flask-script = "^2.0.6"
flask-blueprint = "^1.3.0"
webargs = "^6.1.1"
Flask-Migrate = "^2.5.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
